Renamed Function to NativeFunction
