+++ CalcExpr/Parsing/ParseMatchFunctions.cs
@@ -2,6 +2,7 @@
 using CalcExpr.Expressions;
 using CalcExpr.Expressions.Collections;
 using CalcExpr.Expressions.Components;
+using CalcExpr.Expressions.Functions;
 using CalcExpr.Parsing.Rules;
 using System.Text.RegularExpressions;
 using System.Reflection;
