+++ TestCalcExpr/Expressions/TestFunction.cs
@@ -1,5 +1,6 @@
 ﻿using CalcExpr.Expressions;
 using CalcExpr.Expressions.Components;
+using CalcExpr.Expressions.Functions;
 using System.Linq.Expressions;
 using System.Reflection;
 using TestCalcExpr.Extensions;
@@ -20,14 +21,14 @@ public class TestFunction
         {
             Delegate del = method.ToDelegate();
             IParameter[] parameters = [.. method.GetParameters().ToParameters([])];
-            Function function = new Function(del);
+            NativeFunction function = new NativeFunction(del);
 
             Assert.AreEqual(method, function.Body.Method);
 
             for (int i = 0; i < parameters.Length; i++)
                 Assert.AreEqual(parameters[i], function.Parameters[i]);
 
-            function = new Function(parameters, del);
+            function = new NativeFunction(parameters, del);
             Assert.AreEqual(method, function.Body.Method);
 
             for (int i = 0; i < parameters.Length; i++)
